# Backend/docs/openapi-final.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIGGAB API",
    "version": "2.1.0",
    "description": "Sistema Gestor de Ganado Bovino (SIGGAB) - API REST.\n\nTodos los endpoints (excepto login y refresh) requieren Bearer token.\nRespuesta estandar: { success, data, message, errors }.",
    "contact": {
      "name": "SIGGAB Dev Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenido en POST /auth/login"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "nullable": true
          },
          "message": {
            "type": "string",
            "example": "Operacion exitosa"
          },
          "errors": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "success",
          "data",
          "message",
          "errors"
        ]
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/eventos-sanitarios/tipos": {
      "get": {
        "tags": [
          "EventosSanitarios"
        ],
        "summary": "Listar tipos de evento sanitario",
        "responses": {
          "200": {
            "description": "Catalogo de tipos sanitarios"
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
    "/eventos-sanitarios": {
      "get": {
        "tags": [
          "EventosSanitarios"
        ],
        "summary": "Listar eventos sanitarios",
        "parameters": [
          {
            "in": "query",
            "name": "animal",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "tipo",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "APROBADO",
                "RECHAZADO"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Eventos sanitarios obtenidos"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "tags": [
          "EventosSanitarios"
        ],
        "summary": "Registrar evento sanitario",
        "description": "El registro queda en estado PENDIENTE hasta su autorizacion oficial.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idAnimal",
                  "idTipoEvento",
                  "fechaEvento"
                ],
                "properties": {
                  "idAnimal": {
                    "type": "integer"
                  },
                  "idTipoEvento": {
                    "type": "integer"
                  },
                  "fechaEvento": {
                    "type": "string",
                    "format": "date"
                  },
                  "diagnostico": {
                    "type": "string"
                  },
                  "medicamento": {
                    "type": "string"
                  },
                  "dosis": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Evento sanitario creado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      }
    },
    "/eventos-sanitarios/{id}": {
      "get": {
        "tags": [
          "EventosSanitarios"
        ],
        "summary": "Obtener evento sanitario por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      },
      "patch": {
        "tags": [
          "EventosSanitarios"
        ],
        "summary": "Modificar evento sanitario",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "idTipoEvento": {
                    "type": "integer"
                  },
                  "fechaEvento": {
                    "type": "string",
                    "format": "date"
                  },
                  "diagnostico": {
                    "type": "string"
                  },
                  "medicamento": {
                    "type": "string"
                  },
                  "dosis": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento modificado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      }
    },
    "/eventos-sanitarios/{id}/aprobar": {
      "patch": {
        "tags": [
          "EventosSanitarios"
        ],
        "summary": "Autorizar o rechazar evento sanitario (RF11)",
        "description": "Autorizacion oficial permitida solo para Medico Veterinario.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "estadoAprobacion"
                ],
                "properties": {
                  "estadoAprobacion": {
                    "type": "string",
                    "enum": [
                      "APROBADO",
                      "RECHAZADO"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evento autorizado o rechazado"
          },
          "400": {
            "description": "Error de validacion o regla de negocio"
          },
          "403": {
            "description": "Sin permisos"
          },
          "404": {
            "description": "Evento no encontrado"
          }
        }
      }
    },
...
    "/lotes-productivos": {
      "get": {
        "tags": [
          "ProductivoLotes"
        ],
        "summary": "Listar lotes productivos",
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "APROBADO",
                "RECHAZADO"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lotes obtenidos"
          },
          "401": {
            "description": "No autenticado"
          },
          "403": {
            "description": "Sin permisos"
          }
        }
      },
      "post": {
        "tags": [
          "ProductivoLotes"
        ],
        "summary": "Crear lote productivo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fechaInicio",
                  "fechaFin"
                ],
                "properties": {
                  "fechaInicio": {
                    "type": "string",
                    "format": "date"
                  },
                  "fechaFin": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lote creado"
          },
          "400": {
            "description": "Error de validacion"
          }
        }
      }
    },
    "/lotes-productivos/{id}": {
      "get": {
        "tags": [
          "ProductivoLotes"
        ],
        "summary": "Obtener lote por ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lote obtenido"
          },
          "404": {
            "description": "Lote no encontrado"
          }
        }
      }
    },
    "/lotes-productivos/{id}/validar": {
      "patch": {
        "tags": [
          "ProductivoLotes"
        ],
        "summary": "Validar lote productivo",
        "description": "Solo Administrador puede aprobar o rechazar lotes.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "estado"
                ],
                "properties": {
                  "estado": {
                    "type": "string",
                    "enum": [
                      "APROBADO",
                      "RECHAZADO"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lote validado"
          },
          "400": {
            "description": "Error de validacion"
          },
          "404": {
            "description": "Lote no encontrado"
          }
        }
      }
    },
...
    "/reportes/sanitario": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Generar reporte sanitario",
        "parameters": [
          {
            "in": "query",
            "name": "fechaInicio",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "fechaFin",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "estadoAprobacion",
            "schema": {
              "type": "string",
              "enum": [
                "PENDIENTE",
                "APROBADO",
                "RECHAZADO"
              ]
            }
          },
          {
            "in": "query",
            "name": "idTipoEvento",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte generado (JSON, CSV o PDF)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/reportes/productivo": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "summary": "Generar reporte productivo",
        "parameters": [
          {
            "in": "query",
            "name": "fechaInicio",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "fechaFin",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "idAnimal",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "idLote",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "schema": {
              "type": "string",
              "enum": [
                "json",
                "csv",
                "pdf"
              ],
              "default": "json"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reporte generado (JSON, CSV o PDF)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              },
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}